Added Authentication and CORS

+++ frontend/src/components/ServiceList.js
@@ -1,14 +1,24 @@
 import { useServiceContext } from '../hooks/useServiceContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const ServiceList = ({ service }) => {
     const { dispatch } = useServiceContext()
-    
+    const { user } = useAuthContext()
+
     const handleClick = async () => {
-        const response = await fetch('/api/services/' + service._id, {
-            method: 'DELETE'
+
+        if(!user){
+            return
+        }
+
+        const response = await fetch('http://localhost:4000/api/services/' + service._id, {
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
